refactor(PlacementInputMapper): 패키지 수정
